@@ -3,13 +3,12 @@
 Specifically, the Mann eddy lifetime function and the von Karman energy spectrum.
 """
 
-__all__ = ["VKEnergySpectrum", "MannEddyLifetime", "Mann_linear_exponential_approx"]
+__all__ = ["VKEnergySpectrum", "VKLike_EnergySpectrum", "MannEddyLifetime", "Mann_linear_exponential_approx"]
 
 import io
 import pickle
 from dataclasses import astuple
 from pathlib import Path
-from typing import Union
 
 import matplotlib.pyplot as plt
 import numpy as np
@@ -35,10 +34,69 @@ def VKEnergySpectrum(kL: torch.Tensor) -> torch.Tensor:
     torch.Tensor
         Result of the evaluation
     """
+    # TODO: is this a bug here? since we introduce extra factors of L
+    # NOTE: Originally, this is
+    #      k^(-5/3) (kL / (1 + (kL)^2)^(1/2))^(17/3)
+    #    = k^(-5/3) (kL)^(17/3) / (1 + (kL)^2)^(17/6))
+    # .    -- EXTRA FACTORS OF L INTRODUCED HERE
+    # .  = k^(12/3) / (1 + (kL)^2)^(17/6))
+    # .  = k^4 / (1 + (kL)^2)^(17/6))
+
     return kL**4 / (1.0 + kL**2) ** (17.0 / 6.0)
 
 
-def MannEddyLifetime(kL: Union[torch.Tensor, np.ndarray]) -> torch.Tensor:
+@torch.jit.script
+def VKLike_EnergySpectrum(kL: torch.Tensor) -> torch.Tensor:
+    r"""Evaluate Von Karman energy spectrum without scaling.
+
+    .. math::
+        \widetilde{E}(\boldsymbol{k}) = \left(\frac{k L}{\left(1+(k L)^2\right)^{1 / 2}}\right)^{17 / 3}.
+
+    Parameters
+    ----------
+    kL : torch.Tensor
+        Scaled wave number domain.
+
+    Returns
+    -------
+    torch.Tensor
+        Result of the evaluation
+    """
+    return kL ** (-5.0 / 3.0) * (kL / torch.sqrt(1.0 + kL**2)) ** (17.0 / 3.0)
+
+
+@torch.jit.script
+def LearnableEnergySpectrum(kL: torch.Tensor, p: torch.Tensor, q: torch.Tensor) -> torch.Tensor:
+    r"""
+    Evaluate the learnable energy spectrum, with adjustable low- and high-k slopes.
+
+    The energy spectrum has the form:
+
+    .. math::
+        E(k) = C \frac{(kL)^{2p}}{(1 + (kL)^2)^q}
+
+    where C and L are learned parameters elsewhere in the model. p and q are learnable parameters which control
+    the low- and high-k slopes respectively. For a traditional von Karman spectrum, we have p=2 and q=17/6, which
+    we use to initialize the learnable parameters.
+
+    Parameters
+    ----------
+    kL : torch.Tensor
+        Scaled wave number.
+    p : torch.Tensor
+        Learnable parameter controlling the low-k slope.
+    q : torch.Tensor
+        Learnable parameter controlling the high-k slope.
+
+    Returns
+    -------
+    torch.Tensor
+        Result of the evaluation.
+    """
+    return (kL ** (2.0 * p)) / ((1.0 + kL**2) ** q)
+
+
+def MannEddyLifetime(kL: torch.Tensor | np.ndarray) -> torch.Tensor:
     r"""Evaluate the full Mann eddy lifetime function.
 
     The full Mann eddy lifetime function has the form
@@ -122,7 +180,7 @@ class CPU_Unpickler(pickle.Unpickler):
             return super().find_class(module, name)
 
 
-def plot_loss_logs(log_file: Union[str, Path]):
+def plot_loss_logs(log_file: str | Path):
     """Plot the loss terms from a TensorBoard log file.
 
     Returns a full plot of all loss terms for a specific training log generated by TensorBoard. This is an auxiliary
