@@ -2,13 +2,14 @@
 
 __all__ = [
     "CalibrationProblem",
-    "OnePointSpectraDataGenerator",
-    "PowerSpectraRDT",
-    "OnePointSpectra",
+    "CustomDataFormatter",
+    "generate_kaimal_spectra",
+    "generate_von_karman_spectra",
     "LossAggregator",
+    "OnePointSpectra",
 ]
 
 from .calibration import CalibrationProblem
-from .data_generator import OnePointSpectraDataGenerator
+from .data_generator import CustomDataFormatter, generate_kaimal_spectra, generate_von_karman_spectra
 from .loss_functions import LossAggregator
 from .one_point_spectra import OnePointSpectra
