@@ -8,8 +8,7 @@ from a pre-trained DRD model, or based on some well-known spectra models. ``DRDM
 provides several utilities for plotting the resulting fields through Plotly, which can be done
 in several contexts as well as utilities for saving to VTK for downstream analysis in, e.g.,
 ParaView.
-
-"""
+"""  # noqa
 
 #######################################################################################
 #   .. centered::
