@@ -1,19 +1,20 @@
-"""The global configuration for the test suite"""
+"""The global configuration for the test suite."""
 
 import pytest
 
 
 def pytest_addoption(parser):
-    parser.addoption(
-        "--runslow", action="store_true", default=False, help="run slow tests"
-    )
+    """Add command line options to pytest."""
+    parser.addoption("--runslow", action="store_true", default=False, help="run slow tests")
 
 
 def pytest_configure(config):
+    """Configure pytest."""
     config.addinivalue_line("markers", "slow: mark test as slow to run")
 
 
 def pytest_collection_modifyitems(config, items):
+    """Modify pytest collection."""
     if config.getoption("--runslow"):
         # --runslow given in cli: do not skip slow tests
         return
