@@ -3,10 +3,9 @@
 from pathlib import Path
 
 import numpy as np
-import pytest
 import torch
 
-from drdmannturb.spectra_fitting import OnePointSpectraDataGenerator
+from drdmannturb.spectra_fitting.data_generator import generate_kaimal_spectra
 
 device = "cpu"
 
@@ -20,14 +19,12 @@ def test_kaimal_mann():
 
     k1 = torch.logspace(-1, 2, 20) / zref
 
-    Data = OnePointSpectraDataGenerator(data_points=k1, zref=zref, ustar=ustar).Data
+    Data = generate_kaimal_spectra(data_points=k1, zref=zref, ustar=ustar)
 
     kaimal_mann_spectra_new = Data[1].to("cpu").numpy()
 
     kaimal_mann_spectra_true = (
-        torch.load(fp / "kaimal_mann_data_raw.pt", map_location=torch.device(device))
-        .to("cpu")
-        .numpy()
+        torch.load(fp / "kaimal_mann_data_raw.pt", map_location=torch.device(device)).to("cpu").numpy()
     )
 
     assert np.allclose(kaimal_mann_spectra_new, kaimal_mann_spectra_true)
