@@ -10,14 +10,16 @@ from drdmannturb.fluctuation_generation.fluctuation_field_generator import Fluct
 
 @pytest.mark.skip(reason="Not implemented")
 def test_generator_class():
+    """Test the FluctuationFieldGenerator class."""
     pass
 
 
 class TestFluctuationFieldGenerator:
+    """Test the FluctuationFieldGenerator class."""
+
     @pytest.fixture
     def basic_params(self):
-        """Common parameters for test initialization"""
-
+        """Create dictionary of common parameters for test initialization."""
         return {
             "friction_velocity": 1.0,
             "reference_height": 100.0,
@@ -27,8 +29,7 @@ class TestFluctuationFieldGenerator:
         }
 
     def test_init_invalid_model(self, basic_params):
-        """Test initialization with invalid model"""
-
+        """Test initialization with invalid model."""
         params = basic_params.copy()
         params = basic_params.copy()
         params.update({"model": "InvalidModel", "length_scale": 100.0, "time_scale": 3.0, "energy_spectrum_scale": 1.7})
