@@ -1,4 +1,9 @@
-"""Basic tests for assessing GPU use of the package. During the initial release, the GPU utilization during training was >=95% throughout training. Changes that drop GPU utilization should be considered regressions to package performance."""
+"""
+Basic tests for assessing GPU use of the package.
+
+During the initial release, the GPU utilization during training was >=95% throughout training. Changes that drop GPU
+utilization should be considered regressions to package performance.
+"""
 
 import pytest
 import torch
@@ -11,7 +16,7 @@ from drdmannturb.parameters import (
     PhysicalParameters,
     ProblemParameters,
 )
-from drdmannturb.spectra_fitting import CalibrationProblem, OnePointSpectraDataGenerator
+from drdmannturb.spectra_fitting import CalibrationProblem, generate_kaimal_spectra
 
 device = "cuda" if torch.cuda.is_available() else "cpu"
 
@@ -32,7 +37,13 @@ domain = torch.logspace(-1, 2, 20)
 
 @pytest.mark.slow
 def test_gpu_utilization_synth_fit():
+    """
+    Test that the GPU utilization is >=95% throughout training.
 
+    This test is a simple check to ensure that the GPU utilization is
+    not too low, which would indicate that the model is not using the
+    GPU effectively.
+    """
     pb = CalibrationProblem(
         nn_params=NNParameters(
             nlayers=2,
@@ -41,27 +52,22 @@ def test_gpu_utilization_synth_fit():
             # Specifying the activations is done similarly.
             activations=[nn.ReLU(), nn.ReLU()],
         ),
-        prob_params=ProblemParameters(
-            nepochs=10, learn_nu=False, eddy_lifetime=EddyLifetimeType.TAUNET
-        ),
-        # Note that we have not activated the first order term, but this can be done by passing a value for ``alpha_pen1``
+        prob_params=ProblemParameters(nepochs=10, learn_nu=False, eddy_lifetime=EddyLifetimeType.TAUNET),
+        # Note that we have not activated the first order term,
+        # but this can be done by passing a value for ``alpha_pen1``
         loss_params=LossParameters(alpha_pen2=1.0, beta_reg=1.0e-5),
-        phys_params=PhysicalParameters(
-            L=L, Gamma=Gamma, sigma=sigma, Uref=Uref, domain=domain
-        ),
+        phys_params=PhysicalParameters(L=L, Gamma=Gamma, sigma=sigma, Uref=Uref, domain=domain),
         logging_directory="runs/synthetic_fit",
         device=device,
     )
 
-    Data = OnePointSpectraDataGenerator(zref=zref, data_points=domain).Data
+    Data = generate_kaimal_spectra(data_points=domain, zref=zref, ustar=Uref)
 
     import importlib.util as util
 
-    optimal_parameters = pb.calibrate(data=Data)
+    pb.calibrate(data=Data)
 
     if torch.cuda.is_available() and util.find_spec("pynvml") is not None:
         assert torch.cuda.utilization() >= 95
     else:
-        raise EnvironmentError(
-            "CUDA must be available in test runner with pynvml installed in the environment."
-        )
+        raise OSError("CUDA must be available in test runner with pynvml installed in the environment.")
